requirements.txt
openai==1.12.0
notion-client
watchdog==3.0.0
python-dotenv==0.21.0
pytesseract==0.3.10
opencv-python
Pillow
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
python-docx==0.8.11
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-Login==0.6.1
Flask-WTF==1.0.1
Werkzeug==2.1.2
pytest==7.1.2
beautifulsoup4==4.12.0
streamlit
requests
tqdm
dateparser
unidecode
filetype
sentence-transformers
qdrant-client
openai
langchain
psutil
watchdog
Pillow